+++ packages/grafana-data/src/dataframe/index.ts
@@ -6,6 +6,5 @@ export * from './processDataFrame';
 export * from './dimensions';
 export * from './ArrayDataFrame';
 export * from './DataFrameJSON';
-export { StreamingDataFrame, StreamingFrameOptions, StreamingFrameAction } from './StreamingDataFrame';
 export * from './frameComparisons';
 export { anySeriesWithTimeField } from './utils';
