+++ packages/grafana-runtime/src/services/live.ts
@@ -1,14 +1,10 @@
 import {
   DataFrame,
   DataQueryError,
-  DataQueryResponseData,
-  isDataFrame,
   LiveChannelAddress,
   LiveChannelEvent,
   LiveChannelPresenceStatus,
   LoadingState,
-  StreamingDataFrame,
-  StreamingFrameOptions,
 } from '@grafana/data';
 import { Observable } from 'rxjs';
 
@@ -17,6 +13,36 @@ import { Observable } from 'rxjs';
  */
 export interface LiveDataFilter {
   fields?: string[];
+}
+
+/**
+ * Indicate if the frame is appened or replace
+ *
+ * @public -- but runtime
+ */
+export enum StreamingFrameAction {
+  Append = 'append',
+  Replace = 'replace',
+}
+
+/**
+ * @alpha
+ */
+export interface StreamingFrameOptions {
+  maxLength: number; // 1000
+  maxDelta: number; // how long to keep things
+  action: StreamingFrameAction; // default will append
+}
+
+/**
+ * @alpha
+ */
+export function getStreamingFrameOptions(opts?: Partial<StreamingFrameOptions>): StreamingFrameOptions {
+  return {
+    maxLength: opts?.maxLength ?? 1000,
+    maxDelta: opts?.maxDelta ?? Infinity,
+    action: opts?.action ?? StreamingFrameAction.Append,
+  };
 }
 
 /**
@@ -30,58 +56,7 @@ export interface LiveDataStreamOptions {
   filter?: LiveDataFilter;
 }
 
-/**
- * @alpha -- experimental
- */
-export enum StreamingResponseDataType {
-  NewValuesSameSchema = 'NewValuesSameSchema',
-  FullFrame = 'FullFrame',
-}
-
-/**
- * @alpha -- experimental
- */
-export type StreamingResponseDataTypeToData = {
-  [StreamingResponseDataType.NewValuesSameSchema]: {
-    values: unknown[][];
-  };
-  [StreamingResponseDataType.FullFrame]: {
-    frame: ReturnType<StreamingDataFrame['serialize']>;
-  };
-};
-
-/**
- * @alpha -- experimental
- */
-export type StreamingResponseData<T = StreamingResponseDataType> = T extends StreamingResponseDataType
-  ? {
-      type: T;
-    } & StreamingResponseDataTypeToData[T]
-  : never;
-
-/**
- * @alpha -- experimental
- */
-export const isStreamingResponseData = <T extends StreamingResponseDataType>(
-  responseData: DataQueryResponseData,
-  type: T
-): responseData is StreamingResponseData<T> => 'type' in responseData && responseData.type === type;
-
-const AllStreamingResponseDataTypes = Object.values(StreamingResponseDataType);
-
-/**
- * @alpha -- experimental
- */
-export const isAnyStreamingResponseData = (
-  responseData: DataQueryResponseData
-): responseData is StreamingResponseData =>
-  'type' in responseData && AllStreamingResponseDataTypes.includes(responseData.type);
-
-/**
- * @alpha -- experimental
- */
-export const isStreamingDataFrame = (data: DataQueryResponseData): data is StreamingDataFrame =>
-  isDataFrame(data) && 'packetInfo' in data;
+type StreamingResponseData = any; // ????
 
 /**
  * @alpha -- experimental
