+++ public/app/features/query/state/PanelQueryRunner.ts
@@ -4,7 +4,7 @@ import { MonoTypeOperatorFunction, Observable, of, ReplaySubject, Unsubscribable
 import { map, mergeMap } from 'rxjs/operators';
 
 // Services & Utils
-import { getTemplateSrv, isStreamingDataFrame } from '@grafana/runtime';
+import { getTemplateSrv } from '@grafana/runtime';
 import { getDatasourceSrv } from 'app/features/plugins/datasource_srv';
 import { preProcessPanelData, runRequest } from './runRequest';
 import { isSharedDashboardQuery, runSharedRequest } from '../../../plugins/datasource/dashboard';
@@ -28,7 +28,6 @@ import {
   PanelData,
   rangeUtil,
   ScopedVars,
-  StreamingDataFrame,
   TimeRange,
   TimeZone,
   toDataFrame,
@@ -37,6 +36,8 @@ import {
 import { getDashboardQueryRunner } from './DashboardQueryRunner/DashboardQueryRunner';
 import { mergePanelAndDashData } from './mergePanelAndDashData';
 import { PanelModel } from '../../dashboard/state';
+import { isStreamingDataFrame } from 'app/features/live/data/utils';
+import { StreamingDataFrame } from 'app/features/live/data/StreamingDataFrame';
 
 export interface QueryRunnerOptions<
   TQuery extends DataQuery = DataQuery,
