+++ public/app/plugins/datasource/testdata/runStreams.ts
@@ -9,7 +9,6 @@ import {
   CSVReader,
   Field,
   LoadingState,
-  StreamingDataFrame,
   DataFrameSchema,
   DataFrameData,
 } from '@grafana/data';
@@ -17,6 +16,7 @@ import {
 import { TestDataQuery, StreamingQuery } from './types';
 import { getRandomLine } from './LogIpsum';
 import { liveTimer } from 'app/features/dashboard/dashgrid/liveTimer';
+import { StreamingDataFrame } from 'app/features/live/data/StreamingDataFrame';
 
 export const defaultStreamQuery: StreamingQuery = {
   type: 'signal',
